Switch default strategy.
